Avoid oid[] type in generated query
pgtype doesn't register support OID arrays by default.

+++ internal/pg/query.sql
@@ -1,6 +1,6 @@
 -- name: FindEnumTypes :many
 WITH enums AS (
-  SELECT enumtypid                                       AS enum_type,
+  SELECT enumtypid::int8                                   AS enum_type,
          -- pg_enum row identifier.
          -- The OIDs for pg_enum rows follow a special rule: even-numbered OIDs
          -- are guaranteed to be ordered in the same way as the sort ordering of
@@ -10,10 +10,10 @@ WITH enums AS (
          -- This rule allows the enum comparison routines to avoid catalog
          -- lookups in many common cases. The routines that create and alter enum
          -- types attempt to assign even OIDs to enum values whenever possible.
-         array_agg(oid ORDER BY enumsortorder)           AS enum_oids,
+         array_agg(oid::int8 ORDER BY enumsortorder)       AS enum_oids,
          -- The sort position of this enum value within its enum type. Starts as
          -- 1..n but can be fractional or negative.
-         array_agg(enumsortorder ORDER BY enumsortorder) AS enum_orders,
+         array_agg(enumsortorder ORDER BY enumsortorder)   AS enum_orders,
          -- The textual label for this enum value
          array_agg(enumlabel::text ORDER BY enumsortorder) AS enum_labels
   FROM pg_enum
